% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{aic}
\alias{aic}
\alias{AIC.starmagarch}
\alias{BIC.starmagarch}
\title{Information criterions}
\usage{
\method{AIC}{starmagarch}(object)

\method{BIC}{starmagarch}(object)
}
\arguments{
\item{object}{\code{starmagarch} object}
}
\value{
\code{AIC}: AIC of fitted model.

\code{BIC}: BIC of fitted model.
}
\description{
Akaike's and Bayesian information criterion of a fitted starmagarch model.
}
