% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{correct.names}
\alias{correct.names}
\title{Function for fixing parameter names}
\usage{
correct.names(names)
}
\arguments{
\item{names}{vector of names which are non-unique}
}
\value{
Vector with correct names
}
\description{
Correct the parameter names.
}
