% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{fitSTARMAGARCH}
\alias{fitSTARMAGARCH}
\title{Optimize likelihood using nlminb}
\usage{
fitSTARMAGARCH(
  f,
  data = NULL,
  print = TRUE,
  lower = NULL,
  upper = NULL,
  simple = FALSE
)
}
\arguments{
\item{f}{object produced by CreateLikelihood}

\item{data}{Data matrix (spatial row index, temporal col index)}

\item{print}{logical, indicator for printing summary}

\item{lower}{lower limits for parameters esimates}

\item{upper}{upper limits for parameter estimates}

\item{simple}{logical switch for turning on only returning parameter estimates and sd.}
}
\value{
\code{starmagarch} object.
}
\description{
Optimize likelihood using nlminb
}
