% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{genfunctions}
\alias{genfunctions}
\alias{coef.starmagarch}
\alias{print.starmagarch}
\alias{sigma.starmagarch}
\alias{fittedgarch.starmagarch}
\alias{fitted.starmagarch}
\alias{residuals.starmagarch}
\alias{summary.starmagarch}
\title{Methods for \code{starmagarch} objects}
\usage{
\method{coef}{starmagarch}(object, ...)

\method{print}{starmagarch}(object, ...)

\method{sigma}{starmagarch}(object)

\method{fittedgarch}{starmagarch}(object, ...)

\method{fitted}{starmagarch}(object, ...)

\method{residuals}{starmagarch}(object, ...)

\method{summary}{starmagarch}(object, ...)
}
\arguments{
\item{object}{\code{starmagarch} object}

\item{...}{optionally more fitted model objects.}
}
\value{
\code{coef}: Coefficients of fitted model.

Printout

\code{sigma}: Fitted sigma process, \eqn{\{ \sigma_t(u) \}}.

\code{fittedgarch}: Extract ARMA-residuals (the garch process): \eqn{\widehat y_t(u)}

\code{fitted}: Fitted values of the ARMA process.

\code{residuals}:  Extract standardized residuals of fitted model:
\eqn{z_t(u) = x_t(u) / \sigma_t(u)}

\code{summary}: Summary statistics of fitted model.
}
\description{
Collection of generic functions for \code{starmagarch} objects.
}
