% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{parameterlist2maptemplate}
\alias{parameterlist2maptemplate}
\title{Generate a valid map from parameter list}
\usage{
parameterlist2maptemplate(parameters)
}
\arguments{
\item{parameters}{Parameter list}
}
\value{
List of factors with same structure as \code{parameters}.
}
\description{
Generate a valid map list from parameter list. The map should be used to
specify fixed parameters and which parameters should be set equal. A fixed
parameter is set by setting the corresponding map element to \code{NA} while
two elements in the \code{parameters} list which share the same factor in the
map list are set to be equal.
}
