% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{parvector2list}
\alias{parvector2list}
\title{Function for generating a valid list of parameters from a vector}
\usage{
parvector2list(par, ar = c(2, 1), ma = c(2, 1), arch = c(2, 1),
  garch = c(2, 1))
}
\arguments{
\item{par}{Parameter vector (numeric)}

\item{ar}{Integer of length 2 (spatial, temporal): Spatio-temporal order of AR part}

\item{ma}{Integer of length 2 (spatial, temporal): Spatio-temporal order of MA part}

\item{arch}{Integer of length 2 (spatial, temporal): Spatio-temporal order of ARCH part}

\item{garch}{Integer of length 2 (spatial, temporal): Spatio-temporal order of GARCH part}
}
\description{
Function for generating a valid list of parameters from a vector
}
