% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{CreateLikelihood}
\alias{CreateLikelihood}
\title{Function for estimating circular ARMAGARCH}
\usage{
CreateLikelihood(data, W = NULL, init = apply(data, 1, stats::var),
  parameters = NULL, map = NULL, silent = TRUE)
}
\arguments{
\item{data}{Data matrix (spatial row index, temporal col index)}

\item{W}{Neighbourhood 3D-array}

\item{init}{Initial value of sigma}

\item{parameters}{List of initial parameters}

\item{map}{List of same size as parameters, determining which parameters should be fixed.}

\item{silent}{Logical, indicating wheter optimization print-out should be printet.}
}
\value{
List containing automatic differentiated likelihood function.
}
\description{
Wrapper function for creating a likelihood object from the TMB
package.
}
\examples{
# Parameter list:
parameters <- list(mu    = 5,
                   phi   = matrix(c(.8, .1), ncol = 1),
                   theta = matrix(c(-.42, -.2), ncol = 1),
                   omega = .9,
                   alpha = matrix(c(.1,.05,.2,.1), ncol = 2),
                   beta  = matrix(c(.6,.2), ncol = 1))
m <- c(20, 1)
burnin <- 150
n <- 500

# Simulate data:
set.seed(1234)
W <- create.neighbourhood.array(m = m, sp = 3, type = "queen", torus = TRUE)
y <- simSTARMAGARCH(parameters, n = n, m = m, W = W, burnin = burnin)
# Create likelihood object:
f <- CreateLikelihood(y, W = W, init = apply(y, 1, var),
                      parameters = parameters, silent = FALSE)

# Optimize the likelihood using:
fit <- fitSTARMAGARCH(f, data = y, print = FALSE)
summary(fit)

plot(fit)
plot_garch(fit)
# Checking residuals:
qqnorm(residuals(fit), pch = 19)
qqline(residuals(fit), col = "red")

# If you want to set a parameter to a fixed value,
# for instance the mean "mu" to zero, use the map argument:
parameters$mu = 0
f <- CreateLikelihood(y - 5, W = W,
                      init = apply(y, 1, var),
                      parameters = parameters,
                      map = list(mu = factor(NA)))

# Optimize the likelihood using:
fit <- fitSTARMAGARCH(f, data= y, parameters, print = FALSE)
summary(fit)
}
