% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{fitSTARMAGARCH}
\alias{fitSTARMAGARCH}
\title{Optimize likelihood using nlminb}
\usage{
fitSTARMAGARCH(f, data = NULL, print = TRUE)
}
\arguments{
\item{f}{object produced by CreateLikelihood}

\item{upper}{upper limits for parameter estimates}

\item{lower}{lower limits for parameters esimates}
}
\value{
\code{starmagarch} object.
}
\description{
Optimize likelihood using nlminb
}
