% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{fitSTARMAGARCH}
\alias{fitSTARMAGARCH}
\title{Optimize likelihood using nlminb}
\usage{
fitSTARMAGARCH(f, data = NULL, print = TRUE, lower = NULL,
  upper = NULL)
}
\arguments{
\item{f}{object produced by CreateLikelihood}

\item{data}{Data matrix (spatial row index, temporal col index)}

\item{print}{logical, indicator for printing summary}

\item{lower}{lower limits for parameters esimates}

\item{upper}{upper limits for parameter estimates}
}
\value{
\code{starmagarch} object.
}
\description{
Optimize likelihood using nlminb
}
