% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{plotting}
\alias{plotting}
\alias{plot.starmagarch}
\alias{plot_garch.starmagarch}
\title{Plot \code{starmgarch}}
\usage{
\method{plot}{starmagarch}(x, ...)

\method{plot_garch}{starmagarch}(x, ...)
}
\arguments{
\item{x}{Class \code{starmagarch}}

\item{...}{additional arguments.}
}
\value{
ggplot object

ggplot object
}
\description{
Plot and compare the fitted values to the original data.

Plot and compare the fitted values to the original data.
}
